requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
google-cloud-storage
google-auth
Pillow
pdf2image